+++ src/extension.ts
@@ -1,6 +1,6 @@
 'use strict';
 
-import * as vscode from 'vscode';
+import { window, StatusBarAlignment, commands, workspace, ExtensionContext, ConfigurationTarget, FileType, ProgressLocation, Range, ViewColumn, Uri, TextDocument, Position } from 'vscode';
 import * as jsYaml from 'js-yaml';
 import * as path from 'path';
 import * as fs from 'fs';
@@ -22,37 +22,36 @@ import * as f5FastUtils from './utils/f5FastUtils';
 import * as rpmMgmt from './utils/rpmMgmt';
 import { MgmtClient } from './utils/f5DeviceClient';
 import logger from './utils/logger';
-import { deviceImport } from './cfgImport';
+import { deviceImport, deviceImportOnLoad } from './deviceImport';
 
 import { TextDocumentView } from './editorViews/editorView';
 import { getMiniUcs, makeExplosion } from './cfgExplorer';
-import { window } from 'vscode';
 
 
-export function activate(context: vscode.ExtensionContext) {
+export function activate(context: ExtensionContext) {
 
-	// assign context to global
+	// assign context to global name space
 	ext.context = context;
 
-	// Create a status bar item
-	ext.hostStatusBar = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 15);
+	// Create a status bar items - // todo: move this to f5DeviceClient.ts
+	ext.hostStatusBar = window.createStatusBarItem(StatusBarAlignment.Left, 15);
 	context.subscriptions.push(ext.hostStatusBar);
-	ext.hostNameBar = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 14);
+	ext.hostNameBar = window.createStatusBarItem(StatusBarAlignment.Left, 14);
 	context.subscriptions.push(ext.hostNameBar);
-	ext.fastBar = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 13);
+	ext.fastBar = window.createStatusBarItem(StatusBarAlignment.Left, 13);
 	context.subscriptions.push(ext.fastBar);
-	ext.as3Bar = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 12);
+	ext.as3Bar = window.createStatusBarItem(StatusBarAlignment.Left, 12);
 	context.subscriptions.push(ext.as3Bar);
-	ext.doBar = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 11);
+	ext.doBar = window.createStatusBarItem(StatusBarAlignment.Left, 11);
 	context.subscriptions.push(ext.doBar);
-	ext.tsBar = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 10);
+	ext.tsBar = window.createStatusBarItem(StatusBarAlignment.Left, 10);
 	context.subscriptions.push(ext.tsBar);
 	
 
-	ext.connectBar = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 9);
+	ext.connectBar = window.createStatusBarItem(StatusBarAlignment.Left, 9);
 	context.subscriptions.push(ext.connectBar);
 	ext.connectBar.command = 'f5.connectDevice';
-    ext.connectBar.text = 'F5-FAST -> Connect!';
+    ext.connectBar.text = 'F5 -> Connect!';
 	ext.connectBar.tooltip = 'Click to connect!';
 	ext.connectBar.show();
 
@@ -88,10 +87,10 @@ export function activate(context: vscode.ExtensionContext) {
 	
 	
 	const hostsTreeProvider = new F5TreeProvider('');
-	vscode.window.registerTreeDataProvider('f5Hosts', hostsTreeProvider);
-	vscode.commands.registerCommand('f5.refreshHostsTree', () => hostsTreeProvider.refresh());
+	window.registerTreeDataProvider('f5Hosts', hostsTreeProvider);
+	commands.registerCommand('f5.refreshHostsTree', () => hostsTreeProvider.refresh());
 	
-	context.subscriptions.push(vscode.commands.registerCommand('f5.connectDevice', async (device) => {
+	context.subscriptions.push(commands.registerCommand('f5.connectDevice', async (device) => {
 
 	logger.info('selected device', device);  // preferred at the moment
 
@@ -105,7 +104,7 @@ export function activate(context: vscode.ExtensionContext) {
 		};
 		
 		if (!device) {
-			const bigipHosts: Array<devObj> | undefined = await vscode.workspace.getConfiguration().get('f5.hosts');
+			const bigipHosts: Array<devObj> | undefined = await workspace.getConfiguration().get('f5.hosts');
 
 			if (bigipHosts === undefined) {
 				throw new Error('no hosts in configuration');
@@ -119,7 +118,7 @@ export function activate(context: vscode.ExtensionContext) {
 				return { label: item.device, target: item };
 			});
 
-			device = await vscode.window.showQuickPick(qPickHostList, {placeHolder: 'Select Device'});
+			device = await window.showQuickPick(qPickHostList, {placeHolder: 'Select Device'});
 			if (!device) {
 				throw new Error('user exited device input');
 			} else {
@@ -145,26 +144,19 @@ export function activate(context: vscode.ExtensionContext) {
 		logger.debug(`F5 Connect Discovered ${JSON.stringify(connect)}`);
 		setTimeout( () => { tclTreeProvider.refresh();}, 300);
 
-		/**
-		 * setup explode on connect
-		 * 
-		 */
-		// vscode.commands.executeCommand('f5.cfgExploreOnConnect');
-
-
 	}));
 	
-	context.subscriptions.push(vscode.commands.registerCommand('f5.getProvider', async () => {
+	context.subscriptions.push(commands.registerCommand('f5.getProvider', async () => {
 		const resp: any = await ext.mgmtClient?.makeRequest('/mgmt/tm/auth/source');
 		ext.panel.render(resp);
 	}));
 
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5.getF5HostInfo', async () => {
+	context.subscriptions.push(commands.registerCommand('f5.getF5HostInfo', async () => {
 		var device: string | undefined = ext.hostStatusBar.text;
 		
 		if (!device) {
-			device = await vscode.commands.executeCommand('f5.connectDevice');
+			device = await commands.executeCommand('f5.connectDevice');
 		}
 		
 		if (device === undefined) {
@@ -175,7 +167,7 @@ export function activate(context: vscode.ExtensionContext) {
 		ext.panel.render(resp);
 	}));
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5.disconnect', () => {
+	context.subscriptions.push(commands.registerCommand('f5.disconnect', () => {
 
 		if(ext.mgmtClient) {
 			ext.mgmtClient.disconnect();
@@ -184,27 +176,27 @@ export function activate(context: vscode.ExtensionContext) {
 		}
 	}));
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5.clearPassword', async (item) => {
+	context.subscriptions.push(commands.registerCommand('f5.clearPassword', async (item) => {
 		return hostsTreeProvider.clearPassword(item?.label);
 	}));
 
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5.addHost', async (newHost) => {
+	context.subscriptions.push(commands.registerCommand('f5.addHost', async (newHost) => {
 		return await hostsTreeProvider.addDevice(newHost);
 	}));
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5.removeHost', async (hostID) => {
+	context.subscriptions.push(commands.registerCommand('f5.removeHost', async (hostID) => {
 		return await hostsTreeProvider.removeDevice(hostID);
 	}));
 	
-	context.subscriptions.push(vscode.commands.registerCommand('f5.editHost', async (hostID) => {
+	context.subscriptions.push(commands.registerCommand('f5.editHost', async (hostID) => {
 		
 		logger.debug(`Edit Host command: ${JSON.stringify(hostID)}`);
 		
-		let bigipHosts: {device: string} [] | undefined= vscode.workspace.getConfiguration().get('f5.hosts');
+		let bigipHosts: {device: string} [] | undefined= workspace.getConfiguration().get('f5.hosts');
 		logger.debug(`Current bigipHosts: ${JSON.stringify(bigipHosts)}`);
 		
-		vscode.window.showInputBox({
+		window.showInputBox({
 			prompt: 'Update Device/BIG-IP/Host', 
 			value: hostID.label,
 			ignoreFocusOut: true
@@ -228,11 +220,11 @@ export function activate(context: vscode.ExtensionContext) {
 					}
 				});
 				
-				vscode.workspace.getConfiguration().update('f5.hosts', bigipHosts, vscode.ConfigurationTarget.Global);
+				workspace.getConfiguration().update('f5.hosts', bigipHosts, ConfigurationTarget.Global);
 				setTimeout( () => { hostsTreeProvider.refresh();}, 300);
 			} else {
 		
-				vscode.window.showErrorMessage('Already exists or invalid format: <user>@<host/ip>:<port>');
+				window.showErrorMessage('Already exists or invalid format: <user>@<host/ip>:<port>');
 			}
 		});
 		
@@ -240,9 +232,9 @@ export function activate(context: vscode.ExtensionContext) {
 
 
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5.editDeviceProvider', async (hostID) => {
+	context.subscriptions.push(commands.registerCommand('f5.editDeviceProvider', async (hostID) => {
 		
-		let bigipHosts: {device: string} [] | undefined= vscode.workspace.getConfiguration().get('f5.hosts');
+		let bigipHosts: {device: string} [] | undefined= workspace.getConfiguration().get('f5.hosts');
 
 		const providerOptions: string[] = [
 			'local',
@@ -255,7 +247,7 @@ export function activate(context: vscode.ExtensionContext) {
 			'custom for bigiq'
 		];
 
-		vscode.window.showQuickPick(providerOptions, {placeHolder: 'Default BIGIP providers'})
+		window.showQuickPick(providerOptions, {placeHolder: 'Default BIGIP providers'})
 		.then( async input => {
 
 			logger.debug('user input', input);
@@ -265,7 +257,7 @@ export function activate(context: vscode.ExtensionContext) {
 			}
 
 			if (input === 'custom for bigiq') {
-				input = await vscode.window.showInputBox({
+				input = await window.showInputBox({
 					prompt: "Input custom bigiq login provider"
 				});
 			}
@@ -276,7 +268,7 @@ export function activate(context: vscode.ExtensionContext) {
 				}
 			});
 			
-			vscode.workspace.getConfiguration().update('f5.hosts', bigipHosts, vscode.ConfigurationTarget.Global);
+			workspace.getConfiguration().update('f5.hosts', bigipHosts, ConfigurationTarget.Global);
 
 			setTimeout( () => { hostsTreeProvider.refresh();}, 300);
 		});
@@ -284,9 +276,33 @@ export function activate(context: vscode.ExtensionContext) {
 	}));
 
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5.openSettings', () => {
+	context.subscriptions.push(commands.registerCommand('f5.deviceImport', async (item) => {
+
+
+
+		
+		
+		console.log('asAbsolutePath - ', context.asAbsolutePath);
+		console.log('extensionMode - ', context.extensionMode);
+		console.log('extensionPath - ', context.extensionPath);
+		// console.log('extensionPath - ', context.globalState);
+
+
+		// commands.executeCommand('f5.cfgExploreOnConnect');
+
+
+
+
+		await deviceImport('somethign we got from editor...');
+
+
+	}));
+
+
+
+	context.subscriptions.push(commands.registerCommand('f5.openSettings', () => {
 		//	open settings window and bring the user to the F5 section
-		return vscode.commands.executeCommand("workbench.action.openSettings", "f5");
+		return commands.executeCommand("workbench.action.openSettings", "f5");
 	}));
 
 
@@ -304,7 +320,7 @@ export function activate(context: vscode.ExtensionContext) {
 	 * http://patorjk.com/software/taag/#p=display&h=0&f=Letters&t=FAST
 	 */
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5.installRPM', async (selectedRPM) => {
+	context.subscriptions.push(commands.registerCommand('f5.installRPM', async (selectedRPM) => {
 
 
 		if(selectedRPM) {
@@ -331,24 +347,24 @@ export function activate(context: vscode.ExtensionContext) {
 
 	}));
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5.unInstallRPM', async (rpm) => {
+	context.subscriptions.push(commands.registerCommand('f5.unInstallRPM', async (rpm) => {
 		
 		// if no rpm sent in from update command
 		if(!rpm) {
 			// get installed packages
 			const installedRPMs = await rpmMgmt.installedRPMs();
 			// have user select package
-			rpm = await vscode.window.showQuickPick(installedRPMs, {placeHolder: 'select rpm to remove'});
+			rpm = await window.showQuickPick(installedRPMs, {placeHolder: 'select rpm to remove'});
 		} else {
 			// rpm came from rpm update call...
 		}
 
 		if(!rpm) {	// return error pop-up if quickPick escaped
-			return vscode.window.showWarningMessage('user exited - did not select rpm to un-install');
+			return window.showWarningMessage('user exited - did not select rpm to un-install');
 		}
 
 		const status = await rpmMgmt.unInstallRpm(rpm);
-		vscode.window.showInformationMessage(`rpm ${rpm} removal ${status}`);
+		window.showInformationMessage(`rpm ${rpm} removal ${status}`);
 		// debugger;
 		
 		// used to pause between uninstalling and installing a new version of the same atc
@@ -375,16 +391,16 @@ export function activate(context: vscode.ExtensionContext) {
 
 
 	const tclTreeProvider = new TclTreeProvider();
-	vscode.window.registerTreeDataProvider('as3Tasks', tclTreeProvider);
-	vscode.commands.registerCommand('f5.refreshTclTree', () => tclTreeProvider.refresh());
+	window.registerTreeDataProvider('as3Tasks', tclTreeProvider);
+	commands.registerCommand('f5.refreshTclTree', () => tclTreeProvider.refresh());
 	
 
 	// --- IRULE COMMANDS ---
-	context.subscriptions.push(vscode.commands.registerCommand('f5-tcl.getRule', async (rule) => {
+	context.subscriptions.push(commands.registerCommand('f5-tcl.getRule', async (rule) => {
 		return tclTreeProvider.displayRule(rule);
 	}));
 	
-	context.subscriptions.push(vscode.commands.registerCommand('f5-tcl.deleteRule', async (rule) => {
+	context.subscriptions.push(commands.registerCommand('f5-tcl.deleteRule', async (rule) => {
 		return tclTreeProvider.deleteRule(rule);
 	}));
 
@@ -393,25 +409,25 @@ export function activate(context: vscode.ExtensionContext) {
 	
 	
 	// --- IAPP COMMANDS ---
-	context.subscriptions.push(vscode.commands.registerCommand('f5-tcl.getApp', async (item) => {
+	context.subscriptions.push(commands.registerCommand('f5-tcl.getApp', async (item) => {
 		logger.debug('f5-tcl.getApp command: ', item);
 		return ext.panel.render(item);
 	}));
 
 	
-	context.subscriptions.push(vscode.commands.registerCommand('f5-tcl.getTemplate', async (item) => {
+	context.subscriptions.push(commands.registerCommand('f5-tcl.getTemplate', async (item) => {
 		// returns json view of iApp Template
 		return ext.panel.render(item);
 	}));
 	
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5-tcl.getTMPL', async (item) => {
+	context.subscriptions.push(commands.registerCommand('f5-tcl.getTMPL', async (item) => {
 		// gets the original .tmpl output
 		const temp = await tclTreeProvider.getTMPL(item);
 		tclTreeProvider.displayTMPL(temp);
 	}));
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5-tcl.iAppRedeploy', async (item) => {
+	context.subscriptions.push(commands.registerCommand('f5-tcl.iAppRedeploy', async (item) => {
 		const temp = await tclTreeProvider.iAppRedeploy(item);
 		/**
 		 * setup appropriate response
@@ -421,23 +437,23 @@ export function activate(context: vscode.ExtensionContext) {
 		// return utils.displayJsonInEditor(item);
 	}));
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5-tcl.iAppDelete', async (item) => {
+	context.subscriptions.push(commands.registerCommand('f5-tcl.iAppDelete', async (item) => {
 		const temp = await tclTreeProvider.iAppDelete(item);
 		tclTreeProvider.refresh();
 	}));
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5-tcl.postTMPL', async (item) => {
+	context.subscriptions.push(commands.registerCommand('f5-tcl.postTMPL', async (item) => {
 		const resp: any = await tclTreeProvider.postTMPL(item);
-		vscode.window.showInformationMessage(resp);
+		window.showInformationMessage(resp);
 		return resp;
 	}));
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5-tcl.deleteTMPL', async (item) => {
+	context.subscriptions.push(commands.registerCommand('f5-tcl.deleteTMPL', async (item) => {
 		const resp: any = await tclTreeProvider.deleteTMPL(item);
 		return resp;
 	}));
 	
-	context.subscriptions.push(vscode.commands.registerCommand('f5-tcl.mergeTCL', async (item) => {
+	context.subscriptions.push(commands.registerCommand('f5-tcl.mergeTCL', async (item) => {
 		await tclTreeProvider.mergeTCL(item);
 	}));
 
@@ -459,19 +475,19 @@ export function activate(context: vscode.ExtensionContext) {
 	
 	// setting up hosts tree
 	const fastTreeProvider = new FastTemplatesTreeProvider();
-	vscode.window.registerTreeDataProvider('fastView', fastTreeProvider);
-	vscode.commands.registerCommand('f5-fast.refreshTemplates', () => fastTreeProvider.refresh());
+	window.registerTreeDataProvider('fastView', fastTreeProvider);
+	commands.registerCommand('f5-fast.refreshTemplates', () => fastTreeProvider.refresh());
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5-fast.getInfo', async () => {
+	context.subscriptions.push(commands.registerCommand('f5-fast.getInfo', async () => {
 
 		const resp: any = await ext.mgmtClient?.makeRequest(`/mgmt/shared/fast/info`);
 		ext.panel.render(resp);
 	}));
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5-fast.deployApp', async () => {
+	context.subscriptions.push(commands.registerCommand('f5-fast.deployApp', async () => {
 
 		// get editor window
-		var editor = vscode.window.activeTextEditor;
+		var editor = window.activeTextEditor;
 		if (!editor) {	
 			return; // No open text editor
 		}
@@ -489,7 +505,7 @@ export function activate(context: vscode.ExtensionContext) {
 		if(utils.isValidJson(text)){
 			jsonText = JSON.parse(text);
 		} else {
-			vscode.window.showWarningMessage(`Not valid json object`);
+			window.showWarningMessage(`Not valid json object`);
 			return;
 		}
 		
@@ -504,37 +520,37 @@ export function activate(context: vscode.ExtensionContext) {
 	}));
 
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5-fast.getApp', async (tenApp) => {
+	context.subscriptions.push(commands.registerCommand('f5-fast.getApp', async (tenApp) => {
 
 		const resp: any = await ext.mgmtClient?.makeRequest(`/mgmt/shared/fast/applications/${tenApp}`);
 		ext.panel.render(resp);
 	}));
 
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5-fast.getTask', async (taskId) => {
+	context.subscriptions.push(commands.registerCommand('f5-fast.getTask', async (taskId) => {
 
 		const resp: any = await ext.mgmtClient?.makeRequest(`/mgmt/shared/fast/tasks/${taskId}`);
 		ext.panel.render(resp);
 	}));
 
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5-fast.getTemplate', async (template) => {
+	context.subscriptions.push(commands.registerCommand('f5-fast.getTemplate', async (template) => {
 
 		const resp: any = await ext.mgmtClient?.makeRequest(`/mgmt/shared/fast/templates/${template}`);
 		ext.panel.render(resp);
 	}));
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5-fast.getTemplateSets', async (set) => {
+	context.subscriptions.push(commands.registerCommand('f5-fast.getTemplateSets', async (set) => {
 
 		const resp: any = await ext.mgmtClient?.makeRequest(`/mgmt/shared/fast/templatesets/${set}`);
 		ext.panel.render(resp);
 	}));
 
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5-fast.convJson2Mst', async () => {
+	context.subscriptions.push(commands.registerCommand('f5-fast.convJson2Mst', async () => {
 
 		// get editor window
-		var editor = vscode.window.activeTextEditor;
+		var editor = window.activeTextEditor;
 		if (!editor) {	return; // No open text editor
 		}
 
@@ -555,13 +571,13 @@ export function activate(context: vscode.ExtensionContext) {
 			// - if it's an object and it contains "class" key and value should be "Tenant"
 			utils.displayMstInEditor(JSON.parse(text));
 		} else {
-			vscode.window.showWarningMessage(`not valid json object`);
+			window.showWarningMessage(`not valid json object`);
 		}
 
 
 	}));
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5-fast.postTemplate', async (sFile) => {
+	context.subscriptions.push(commands.registerCommand('f5-fast.postTemplate', async (sFile) => {
 
 		let text: string | Buffer;
 
@@ -569,7 +585,7 @@ export function activate(context: vscode.ExtensionContext) {
 			// not right click from explorer view, so gather file details
 
 			// get editor window
-			var editor = vscode.window.activeTextEditor;
+			var editor = window.activeTextEditor;
 			if (!editor) {	
 				return; // No open text editor
 			}
@@ -593,7 +609,7 @@ export function activate(context: vscode.ExtensionContext) {
 		fastTreeProvider.refresh();
 	}));
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5-fast.postTemplateSet', async (sPath) => {
+	context.subscriptions.push(commands.registerCommand('f5-fast.postTemplateSet', async (sPath) => {
 
 		logger.debug('postTemplateSet selection', sPath);
 		let wkspPath;
@@ -603,24 +619,24 @@ export function activate(context: vscode.ExtensionContext) {
 			// didn't get a path passed in from right click, so we have to gather necessary details
 
 			// get list of open workspaces
-			const workspaces = vscode.workspace.workspaceFolders;
+			const workspaces = workspace.workspaceFolders;
 			logger.debug('workspaces', workspaces);
 			
 			// if no open workspace...
 			if(!workspaces) {
 				// Show message to select workspace
-				await vscode.window.showInformationMessage('See top bar to open a workspace with Fast Templates first');
+				await window.showInformationMessage('See top bar to open a workspace with Fast Templates first');
 				// pop up to selecte a workspace
-				await vscode.window.showWorkspaceFolderPick();
+				await window.showWorkspaceFolderPick();
 				// return to begining of function to try again
-				return vscode.commands.executeCommand('f5-fast.postTemplateSet');
+				return commands.executeCommand('f5-fast.postTemplateSet');
 			}
 		
-			const folder1 = vscode.workspace.workspaceFolders![0]!.uri;
+			const folder1 = workspace.workspaceFolders![0]!.uri;
 			wkspPath = folder1.fsPath;
-			const folder2 = await vscode.workspace.fs.readDirectory(folder1);
+			const folder2 = await workspace.fs.readDirectory(folder1);
 		
-			logger.debug('workspace name', vscode.workspace.name);
+			logger.debug('workspace name', workspace.name);
 			
 			/**
 			 * having problems typing the workspaces to a list for quick pick
@@ -629,27 +645,27 @@ export function activate(context: vscode.ExtensionContext) {
 			// let wkspc;
 			// if (workspaces.length > 1) {
 			// 	// if more than one workspace open, have user select the workspace
-			// 	wkspc = await vscode.window.showQuickPick(workspaces);
+			// 	wkspc = await window.showQuickPick(workspaces);
 			// } else {
 			// 	// else select the first workspace
 			// 	wkspc = workspaces[0];
 			// }
 			
 			let wFolders = [];
-			for (const [name, type] of await vscode.workspace.fs.readDirectory(folder1)) {
+			for (const [name, type] of await workspace.fs.readDirectory(folder1)) {
 
-				if (type === vscode.FileType.Directory){
+				if (type === FileType.Directory){
 					logger.debug('---directory', name);
 					wFolders.push(name);
 				}
 			};
 
 			// have user select first level folder in workspace
-			selectedFolder = await vscode.window.showQuickPick(wFolders);
+			selectedFolder = await window.showQuickPick(wFolders);
 			
 			if(!selectedFolder) {
 				// if user "escaped" folder selection window
-				return vscode.window.showInformationMessage('Must select a Fast Template Set folder');
+				return window.showInformationMessage('Must select a Fast Template Set folder');
 			}
 			logger.debug('workspace path', wkspPath);
 			logger.debug('workspace folder', selectedFolder);
@@ -666,7 +682,7 @@ export function activate(context: vscode.ExtensionContext) {
 		fastTreeProvider.refresh();
 	}));
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5-fast.deleteFastApp', async (tenApp) => {
+	context.subscriptions.push(commands.registerCommand('f5-fast.deleteFastApp', async (tenApp) => {
 		
 		// var device: string | undefined = ext.hostStatusBar.text;
 		// const password = await utils.getPassword(device);
@@ -680,11 +696,11 @@ export function activate(context: vscode.ExtensionContext) {
 	}));
 
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5-fast.deleteFastTempSet', async (tempSet) => {
+	context.subscriptions.push(commands.registerCommand('f5-fast.deleteFastTempSet', async (tempSet) => {
 
 		const resp = await f5FastApi.delTempSet(tempSet.label);
 
-		vscode.window.showInformationMessage(`Fast Template Set Delete: ${resp.data.message}`);
+		window.showInformationMessage(`Fast Template Set Delete: ${resp.data.message}`);
 
 		// give a little time to finish
 		await new Promise(resolve => { setTimeout(resolve, 1000); });
@@ -694,7 +710,7 @@ export function activate(context: vscode.ExtensionContext) {
 
 
 	const fastPanel = new FastWebView();
-	context.subscriptions.push(vscode.commands.registerCommand('f5-fast.renderHtmlPreview', async (item) => {
+	context.subscriptions.push(commands.registerCommand('f5-fast.renderHtmlPreview', async (item) => {
 
 		let text: string = 'empty';
 		let title: string = 'Fast Template';	
@@ -723,7 +739,7 @@ export function activate(context: vscode.ExtensionContext) {
 
 		} else {
 			// right-click or commandpalette initiation, so get editor text
-			var editor = vscode.window.activeTextEditor;
+			var editor = window.activeTextEditor;
 			if (editor) {	
 				if (editor?.selection?.isEmpty) {
 					text = editor.document.getText();	// entire editor/doc window
@@ -758,10 +774,10 @@ export function activate(context: vscode.ExtensionContext) {
 	
 	// setting up as3 tree
 	const as3Tree = new AS3TreeProvider();
-	vscode.window.registerTreeDataProvider('as3Tenants', as3Tree);
-	vscode.commands.registerCommand('f5-as3.refreshTenantsTree', () => as3Tree.refresh());
+	window.registerTreeDataProvider('as3Tenants', as3Tree);
+	commands.registerCommand('f5-as3.refreshTenantsTree', () => as3Tree.refresh());
 	
-	context.subscriptions.push(vscode.commands.registerCommand('f5-as3.getDecs', async (tenant) => {
+	context.subscriptions.push(commands.registerCommand('f5-as3.getDecs', async (tenant) => {
 
 		// set blank value if not defined -> get all tenants dec
 		tenant = tenant ? tenant : '';
@@ -771,18 +787,18 @@ export function activate(context: vscode.ExtensionContext) {
 	}));
 
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5-as3.fullTenant', async (tenant) => {
-		vscode.commands.executeCommand('f5-as3.getDecs', `${tenant.label}?show=full`);
+	context.subscriptions.push(commands.registerCommand('f5-as3.fullTenant', async (tenant) => {
+		commands.executeCommand('f5-as3.getDecs', `${tenant.label}?show=full`);
 	}));
-	context.subscriptions.push(vscode.commands.registerCommand('f5-as3.expandedTenant', async (tenant) => {
-		vscode.commands.executeCommand('f5-as3.getDecs', `${tenant.label}?show=expanded`);
+	context.subscriptions.push(commands.registerCommand('f5-as3.expandedTenant', async (tenant) => {
+		commands.executeCommand('f5-as3.getDecs', `${tenant.label}?show=expanded`);
 	}));
 	
 	
-	context.subscriptions.push(vscode.commands.registerCommand('f5-as3.deleteTenant', async (tenant) => {
+	context.subscriptions.push(commands.registerCommand('f5-as3.deleteTenant', async (tenant) => {
 		
-	    const progress = await vscode.window.withProgress({
-			location: vscode.ProgressLocation.Notification,
+	    const progress = await window.withProgress({
+			location: ProgressLocation.Notification,
 			title: `Deleting ${tenant.label} Tenant`
 		}, async (progress) => {
 			
@@ -798,10 +814,10 @@ export function activate(context: vscode.ExtensionContext) {
 		as3Tree.refresh();
 	}));
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5-as3.getTask', (id) => {
+	context.subscriptions.push(commands.registerCommand('f5-as3.getTask', (id) => {
 
-		vscode.window.withProgress({
-			location: vscode.ProgressLocation.Notification,
+		window.withProgress({
+			location: ProgressLocation.Notification,
 			title: `Getting AS3 Task`
 		}, async () => {
 
@@ -811,9 +827,9 @@ export function activate(context: vscode.ExtensionContext) {
 
 	}));
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5-as3.postDec', async () => {
+	context.subscriptions.push(commands.registerCommand('f5-as3.postDec', async () => {
 
-		ext.as3AsyncPost = vscode.workspace.getConfiguration().get('f5.as3Post.async');
+		ext.as3AsyncPost = workspace.getConfiguration().get('f5.as3Post.async');
 
 		let postParam;
 		if(ext.as3AsyncPost) {
@@ -822,7 +838,7 @@ export function activate(context: vscode.ExtensionContext) {
 			postParam = undefined;
 		}
 
-		var editor = vscode.window.activeTextEditor;
+		var editor = window.activeTextEditor;
 		if (!editor) {
 			return; // No open text editor
 		}
@@ -835,7 +851,7 @@ export function activate(context: vscode.ExtensionContext) {
 		} 
 
 		if (!utils.isValidJson(text)) {
-			return vscode.window.showErrorMessage('Not valid JSON object');
+			return window.showErrorMessage('Not valid JSON object');
 		}
 
 		const resp = await f5Api.postAS3Dec(postParam, JSON.parse(text));
@@ -849,11 +865,11 @@ export function activate(context: vscode.ExtensionContext) {
 	 * 		try to figure out if it's as3/do/ts, then apply the appropriate schema reference in the object
 	 * 	if it detects the schema already there, it will remove it.
 	 */
-	context.subscriptions.push(vscode.commands.registerCommand('f5.injectSchemaRef', async () => {
+	context.subscriptions.push(commands.registerCommand('f5.injectSchemaRef', async () => {
 
-		vscode.window.showWarningMessage('experimental feature in development');
+		window.showWarningMessage('experimental feature in development');
 		
-		var editor = vscode.window.activeTextEditor;
+		var editor = window.activeTextEditor;
 		if (!editor) {
 			return; // No open text editor
 		}
@@ -866,7 +882,7 @@ export function activate(context: vscode.ExtensionContext) {
 		} 
 
 		if (!utils.isValidJson(text)) {
-			return vscode.window.showErrorMessage('Not valid JSON object');
+			return window.showErrorMessage('Not valid JSON object');
 		}
 		
 		var newText = JSON.parse(text);
@@ -881,10 +897,10 @@ export function activate(context: vscode.ExtensionContext) {
 			} else if (newText.hasOwnProperty('class') && newText.class === 'Telemetry') {
 				newText['$schema'] = git.latestTSschema;
 			} else {
-				vscode.window.showInformationMessage(`Could not find base declaration class for as3/do/ts`);
+				window.showInformationMessage(`Could not find base declaration class for as3/do/ts`);
 			}
 		} else {
-			vscode.window.showInformationMessage(`Removing ${newText.$schema}`);
+			window.showInformationMessage(`Removing ${newText.$schema}`);
 			delete newText.$schema;
 
 		}
@@ -892,13 +908,13 @@ export function activate(context: vscode.ExtensionContext) {
 		logger.debug(`newText below`);
 		logger.debug(newText);
 
-		const {activeTextEditor} = vscode.window;
+		const {activeTextEditor} = window;
 
         if (activeTextEditor && activeTextEditor.document.languageId === 'json') {
             const {document} = activeTextEditor;
 			const firstLine = document.lineAt(0);
 			const lastLine = document.lineAt(document.lineCount - 1);
-			var textRange = new vscode.Range(0,
+			var textRange = new Range(0,
 			firstLine.range.start.character,
 			document.lineCount - 1,
 			lastLine.range.end.character);
@@ -906,16 +922,16 @@ export function activate(context: vscode.ExtensionContext) {
 				edit.replace(textRange, newText);
 			});
             // if (firstLine.text !== '42') {
-            //     const edit = new vscode.WorkspaceEdit();
+            //     const edit = new WorkspaceEdit();
             //     edit.insert(document.uri, firstLine.range.start, '42\n');
-            //     return vscode.workspace.applyEdit(edit)
+            //     return workspace.applyEdit(edit)
             // }
         }
-		// const { activeTextEditor } = vscode.window;
+		// const { activeTextEditor } = window;
 		// const { document } = activeTextEditor;
 
 		// const fullText = document.getText();
-		// const fullRange = new vscode.Range(
+		// const fullRange = new Range(
 		// 	document.positionAt(0),
 		// 	document.positionAt(fullText.length - 1)
 		// )
@@ -926,13 +942,13 @@ export function activate(context: vscode.ExtensionContext) {
 		
 		// editor.edit(edit => {
 		// 	const startPosition = new Position(0, 0);
-		// 	const endPosition = vscode.TextDocument.lineAt(document.lineCount - 1).range.end;
+		// 	const endPosition = TextDocument.lineAt(document.lineCount - 1).range.end;
 		// 	edit.replace(new Range(startPosition, endPosition), newText);
 		// });
 
 		// var firstLine = textEdit.document.lineAt(0);
 		// var lastLine = textEditor.document.lineAt(textEditor.document.lineCount - 1);
-		// var textRange = new vscode.Range(0,
+		// var textRange = new Range(0,
 		// firstLine.range.start.character,
 		// textEditor.document.lineCount - 1,
 		// lastLine.range.end.character);
@@ -971,15 +987,15 @@ export function activate(context: vscode.ExtensionContext) {
 
 
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5-ts.info', async () => {
+	context.subscriptions.push(commands.registerCommand('f5-ts.info', async () => {
 		const resp: any = await ext.mgmtClient?.makeRequest('/mgmt/shared/telemetry/info');
 		ext.panel.render(resp);
 	}));
 
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5-ts.getDec', async () => {
-		await vscode.window.withProgress({
-			location: vscode.ProgressLocation.Notification,
+	context.subscriptions.push(commands.registerCommand('f5-ts.getDec', async () => {
+		await window.withProgress({
+			location: ProgressLocation.Notification,
 			title: `Getting TS Dec`
 		}, async () => {
 			const resp: any = await ext.mgmtClient?.makeRequest(`/mgmt/shared/telemetry/declare`);
@@ -987,10 +1003,10 @@ export function activate(context: vscode.ExtensionContext) {
 		});
 	}));
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5-ts.postDec', async () => {
+	context.subscriptions.push(commands.registerCommand('f5-ts.postDec', async () => {
 		
 		// if selected text, capture that, if not, capture entire document
-		var editor = vscode.window.activeTextEditor;
+		var editor = window.activeTextEditor;
 		let text: string;
 		if(editor) {
 			if (editor.selection.isEmpty) {
@@ -1000,12 +1016,12 @@ export function activate(context: vscode.ExtensionContext) {
 			} 
 
 			if (!utils.isValidJson(text)) {
-				return vscode.window.showErrorMessage('Not valid JSON object');
+				return window.showErrorMessage('Not valid JSON object');
 			}
 		}
 
-		const progress = await vscode.window.withProgress({
-			location: vscode.ProgressLocation.Notification,
+		const progress = await window.withProgress({
+			location: ProgressLocation.Notification,
 			title: `Posting TS Dec`
 		}, async () => {
 			const resp: any = await ext.mgmtClient?.makeRequest(`/mgmt/shared/telemetry/declare`, {
@@ -1016,7 +1032,7 @@ export function activate(context: vscode.ExtensionContext) {
 		});
 	}));
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5.getGitHubExample', async (decUrl) => {
+	context.subscriptions.push(commands.registerCommand('f5.getGitHubExample', async (decUrl) => {
 
 		const resp = await extAPI.makeRequest({	url: decUrl	});
 		return ext.panel.render(resp);
@@ -1039,10 +1055,10 @@ export function activate(context: vscode.ExtensionContext) {
  * 	http://patorjk.com/software/taag/#p=display&h=0&f=ANSI%20Regular&t=DO
  */
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5-do.getDec', async () => {
+	context.subscriptions.push(commands.registerCommand('f5-do.getDec', async () => {
 
-		await vscode.window.withProgress({
-			location: vscode.ProgressLocation.Notification,
+		await window.withProgress({
+			location: ProgressLocation.Notification,
 			title: `Getting DO Dec`
 		}, async () => {
 			const resp: any = await ext.mgmtClient?.makeRequest(`/mgmt/shared/declarative-onboarding/`);
@@ -1052,9 +1068,9 @@ export function activate(context: vscode.ExtensionContext) {
 
 	}));
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5-do.postDec', async () => {
+	context.subscriptions.push(commands.registerCommand('f5-do.postDec', async () => {
 
-		var editor = vscode.window.activeTextEditor;
+		var editor = window.activeTextEditor;
 		if (!editor) {
 			return; // No open text editor
 		}
@@ -1067,7 +1083,7 @@ export function activate(context: vscode.ExtensionContext) {
 		} 
 
 		if (!utils.isValidJson(text)) {
-			return vscode.window.showErrorMessage('Not valid JSON object');
+			return window.showErrorMessage('Not valid JSON object');
 		}
 
 		const resp = await f5Api.postDoDec(JSON.parse(text));
@@ -1075,10 +1091,10 @@ export function activate(context: vscode.ExtensionContext) {
 	}));
 
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5-do.inspect', async () => {
+	context.subscriptions.push(commands.registerCommand('f5-do.inspect', async () => {
 
-		await vscode.window.withProgress({
-			location: vscode.ProgressLocation.Notification,
+		await window.withProgress({
+			location: ProgressLocation.Notification,
 			title: `Getting DO Inspect`
 		}, async () => {
 			const resp: any = await ext.mgmtClient?.makeRequest(`/mgmt/shared/declarative-onboarding/inspect`);
@@ -1089,10 +1105,10 @@ export function activate(context: vscode.ExtensionContext) {
 
 
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5-do.getTasks', async () => {
+	context.subscriptions.push(commands.registerCommand('f5-do.getTasks', async () => {
 
-		await vscode.window.withProgress({
-			location: vscode.ProgressLocation.Notification,
+		await window.withProgress({
+			location: ProgressLocation.Notification,
 			title: `Getting DO Tasks`
 		}, async () => {
 			const resp: any = await ext.mgmtClient?.makeRequest(`/mgmt/shared/declarative-onboarding/task`);
@@ -1119,7 +1135,7 @@ export function activate(context: vscode.ExtensionContext) {
 
 
 	// register example delarations tree
-	vscode.window.registerTreeDataProvider('decExamples', new ExampleDecsProvider());
+	window.registerTreeDataProvider('decExamples', new ExampleDecsProvider());
 
 
 	// /**
@@ -1135,9 +1151,9 @@ export function activate(context: vscode.ExtensionContext) {
 	//  */
 
 	const cfgProvider = new CfgProvider();
-	vscode.window.registerTreeDataProvider('cfgTree', cfgProvider);
+	window.registerTreeDataProvider('cfgTree', cfgProvider);
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5.cfgExploreOnConnect', async (item) => {
+	context.subscriptions.push(commands.registerCommand('f5.cfgExploreOnConnect', async (item) => {
 
 		/**
 		 * now to ready the archive contents and feed to corkscrew...
@@ -1155,7 +1171,7 @@ export function activate(context: vscode.ExtensionContext) {
              * loop this back into the connect flow, since we have the device, automatically connect
              *  - this should probably happen in the main extension.ts
              */
-            // await vscode.commands.executeCommand('f5.connectDevice', item.label);
+            // await commands.executeCommand('f5.connectDevice', item.label);
             return window.showWarningMessage('Connect to BIGIP Device first');
 		}
 		
@@ -1176,7 +1192,7 @@ export function activate(context: vscode.ExtensionContext) {
 	/**
 	 * this command is exposed via right click in editor so user does not have to connect to F5
 	 */
-	context.subscriptions.push(vscode.commands.registerCommand('f5.cfgExplore', async (item) => {
+	context.subscriptions.push(commands.registerCommand('f5.cfgExplore', async (item) => {
 
 		let expl;
 		if (item._fsPath) {
@@ -1197,11 +1213,11 @@ export function activate(context: vscode.ExtensionContext) {
 
 	}));
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5.cfgExploreClear', async (text) => {
+	context.subscriptions.push(commands.registerCommand('f5.cfgExploreClear', async (text) => {
 		cfgProvider.clear();
 	}));
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5.cfgExplore-show', async (text) => {
+	context.subscriptions.push(commands.registerCommand('f5.cfgExplore-show', async (text) => {
 		cfgProvider.render(text);
 	}));
 
@@ -1217,8 +1233,8 @@ export function activate(context: vscode.ExtensionContext) {
 	//  * 
 	//  */
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5.jsonYmlConvert', async () => {
-		const editor = vscode.window.activeTextEditor;
+	context.subscriptions.push(commands.registerCommand('f5.jsonYmlConvert', async () => {
+		const editor = window.activeTextEditor;
 		if(!editor) {
 			return;
 		}
@@ -1245,9 +1261,9 @@ export function activate(context: vscode.ExtensionContext) {
 	 * refactor the json<->yaml/base64-encode/decode stuff to follow the following logic
 	 * based off of the vscode-extension-examples document-editing-sample
 	 */
-	// let disposable = vscode.commands.registerCommand('extension.reverseWord', function () {
+	// let disposable = commands.registerCommand('extension.reverseWord', function () {
 	// 	// Get the active text editor
-	// 	let editor = vscode.window.activeTextEditor;
+	// 	let editor = window.activeTextEditor;
 
 	// 	if (editor) {
 	// 		let document = editor.document;
@@ -1262,8 +1278,8 @@ export function activate(context: vscode.ExtensionContext) {
 	// 	}
 	// });
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5.b64Encode', () => {
-		const editor = vscode.window.activeTextEditor;
+	context.subscriptions.push(commands.registerCommand('f5.b64Encode', () => {
+		const editor = window.activeTextEditor;
 		if(!editor){
 			return;
 		}
@@ -1275,8 +1291,8 @@ export function activate(context: vscode.ExtensionContext) {
 	}));
 
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5.b64Decode', () => {
-		const editor = vscode.window.activeTextEditor;
+	context.subscriptions.push(commands.registerCommand('f5.b64Decode', () => {
+		const editor = window.activeTextEditor;
 		if(!editor){
 			return;
 		}
@@ -1288,14 +1304,14 @@ export function activate(context: vscode.ExtensionContext) {
 	}));
 
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5.makeRequest', async () => {
+	context.subscriptions.push(commands.registerCommand('f5.makeRequest', async () => {
 		/**
 		 * make open/raw https call
 		 * 
 		 */
 
 		logger.debug('executing f5.makeRequest');
-		const editor = vscode.window.activeTextEditor;
+		const editor = window.activeTextEditor;
 		let resp;
 
 		if(editor){
@@ -1332,8 +1348,8 @@ export function activate(context: vscode.ExtensionContext) {
 
 			if(text.url.includes('http')) {
 
-				resp = await vscode.window.withProgress({
-					location: vscode.ProgressLocation.Notification,
+				resp = await window.withProgress({
+					location: ProgressLocation.Notification,
 					title: `Making External API Request`,
 					cancellable: true
 				}, async (progress, token) => {
@@ -1350,8 +1366,8 @@ export function activate(context: vscode.ExtensionContext) {
 				
 			} else {
 				
-				resp = await vscode.window.withProgress({
-					location: vscode.ProgressLocation.Notification,
+				resp = await window.withProgress({
+					location: ProgressLocation.Notification,
 					title: `Making API Request`,
 					cancellable: true
 				}, async (progress, token) => {
@@ -1364,7 +1380,7 @@ export function activate(context: vscode.ExtensionContext) {
 					// f5 device call
 					if(!ext.mgmtClient) {
 						// connect to f5 if not already connected
-						await vscode.commands.executeCommand('f5.connectDevice');
+						await commands.executeCommand('f5.connectDevice');
 					}
 				
 					logger.debug('device call -> ', JSON.stringify(text));
@@ -1383,9 +1399,9 @@ export function activate(context: vscode.ExtensionContext) {
 	}));
 
 
-	context.subscriptions.push(vscode.commands.registerCommand('f5.remoteCommand', async () => {
+	context.subscriptions.push(commands.registerCommand('f5.remoteCommand', async () => {
 
-		const cmd = await vscode.window.showInputBox({ placeHolder: 'Bash Command to Execute?', ignoreFocusOut: true });
+		const cmd = await window.showInputBox({ placeHolder: 'Bash Command to Execute?', ignoreFocusOut: true });
 		
 		if ( cmd === undefined ) {
 			// maybe just showInformationMessage and exit instead of error?
@@ -1405,49 +1421,50 @@ export function activate(context: vscode.ExtensionContext) {
 
 
 
-	context.subscriptions.push(vscode.commands.registerCommand('chuckJoke', async () => {
+	context.subscriptions.push(commands.registerCommand('chuckJoke', async () => {
 
-		await deviceImport();
 
-		// const newEditorColumn = ext.settings.previewColumn;
-		// const window = vscode.window.visibleTextEditors;
-		// let viewColumn: vscode.ViewColumn | undefined;
+		const newEditorColumn = ext.settings.previewColumn;
+		const wndw = window.visibleTextEditors;
+		let viewColumn: ViewColumn | undefined;
 		
-		// window.forEach(el => {
-		// 	// const el1 = element;
-		// 	if (el.document.fileName === 'chuck-joke.json') {
-		// 		//logger.debug('f5-fast.json editor column', el1.viewColumn);
-		// 		viewColumn = el.viewColumn;
-		// 	}
-		// });
+		wndw.forEach(el => {
+			// const el1 = element;
+			if (el.document.fileName === 'chuck-joke.json') {
+				//logger.debug('f5-fast.json editor column', el1.viewColumn);
+				viewColumn = el.viewColumn;
+			}
+		});
 		
 		
-		// const resp: any = await extAPI.makeRequest({url: 'https://api.chucknorris.io/jokes/random'});
-		// // let activeColumn = vscode.window.activeTextEditor?.viewColumn;
+		const resp: any = await extAPI.makeRequest({url: 'https://api.chucknorris.io/jokes/random'});
+		// let activeColumn = window.activeTextEditor?.viewColumn;
 		
-		// logger.debug('chuck-joke->resp.data', resp.data);
+		logger.debug('chuck-joke->resp.data', resp.data);
 
-		// const content = JSON.stringify(resp.data, undefined, 4);
+		const content = JSON.stringify(resp.data, undefined, 4);
 
-		// // if vClm has a value assign it, else set column 1
-		// viewColumn = viewColumn ? viewColumn : newEditorColumn;
+		// if vClm has a value assign it, else set column 1
+		viewColumn = viewColumn ? viewColumn : newEditorColumn;
 
-		// var vDoc: vscode.Uri = vscode.Uri.parse("untitled:" + "chuck-Joke.json");
-		// vscode.workspace.openTextDocument(vDoc)
-		// .then((a: vscode.TextDocument) => {
-		// 	vscode.window.showTextDocument(a, viewColumn, false).then(e => {
-		// 		e.edit(edit => {
-		// 			const startPosition = new vscode.Position(0, 0);
-		// 			const endPosition = a.lineAt(a.lineCount - 1).range.end;
-		// 			edit.replace(new vscode.Range(startPosition, endPosition), content);
-		// 		});
-		// 	});
-		// });
+		var vDoc: Uri = Uri.parse("untitled:" + "chuck-Joke.json");
+		workspace.openTextDocument(vDoc)
+		.then((a: TextDocument) => {
+			window.showTextDocument(a, viewColumn, false).then(e => {
+				e.edit(edit => {
+					const startPosition = new Position(0, 0);
+					const endPosition = a.lineAt(a.lineCount - 1).range.end;
+					edit.replace(new Range(startPosition, endPosition), content);
+				});
+			});
+		});
 
 
 		// chuckJoke1();
 
 	}));
+
+	deviceImportOnLoad(context.extensionPath);
 
 }